refactor(training): domain/common/enums.py como SSOT dos 12 StrEnums (Fase 5.1)
Extrai os 12 StrEnums do domínio training para módulo dedicado
src/training/domain/common/enums.py como fonte única:

- RoleLabel (antes em domain/rules.py)
- TrainingSessionStatus, SessionBlockPhase, SessionBlockIntensity,
  AttendanceStatus, AttendanceSource, ExecutionType,
  IndividualizationMode, SessionObjectiveOrigin, ConversationOutcome,
  RecommendationActionType, RecommendationStatus, RecommendationPriority
  (antes em domain/entities.py)

entities.py e rules.py convertidos em shims — re-exportam os enums
para preservar a surface pública existente. Identidade de classe
preservada (os `is` continuam verdadeiros através dos imports).

Preparação para Fase 5.3 (split de entities por agregado), onde cada
agregado consumirá os enums diretamente de common/.

Invariantes:
- Suite training/tests: 328 passed, 19 skipped
- test_public_surface.py: PASS
- test_route_inventory.py: PASS
