+++ src/training/domain/common/__init__.py
@@ -1,0 +1,1 @@
+"""Primitivos compartilhados do domínio training (enums, exceções base)."""
